feat: implement idempotent database operations and security improvements
- Switch database writer from insert to upsert with message_id for idempotency
- Add application-level deduplication check to prevent unnecessary operations
- Create database migration with unique constraints for message deduplication
- Add required file_path field to ConversationData model across all instantiations
- Improve WebSocket error handling with specific exception types
- Move WebSocketDisconnect import to module level for better code organization
- Enhance security configuration with reduced JWT expiry and disabled public signup
- Refactor mock classes into dedicated modules for better maintainability
- Fix deprecated asyncio calls and add missing imports in documentation
- Update all test assertions to match new behavior patterns

All tests passing with comprehensive coverage of idempotency features.
